src/py/databricks_apps/simple_lakebase_app/pyproject.toml
[project]
name = "simple-lakebase-app"
version = "0.1.0"
description = "A simple Streamlit app for Lakebase data exploration"
requires-python = ">=3.9"
dependencies = [
    "streamlit==1.39.0",
    "pandas==2.3.1",
    "databricks-sdk==0.57.0",
    "psycopg2-binary==2.9.9",
    "sqlalchemy==2.0.41",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
